重构传输层，添加 InteractsWithCallbacks 特性，更新关闭方法并移除不必要的回调设置
